Move the Ayesha class to its own file for consistent syntax highlighting
